fix: query rules tests

+++ cypress/integration/add-always-trigger-promoteResult-action.spec.js
@@ -2,6 +2,9 @@ import generateName from '../utils/generateName';
 import { base_url, username, password, app_url, cluster } from '../utils/index';
 import { PAGE_LOAD_TIME } from './contants';
 
+// Track query rule id, to delete later
+let ruleId;
+
 describe('Query Rule creation with trigger index and script action', () => {
 	before(() => {
 		cy.window().then((win) => {
@@ -26,12 +29,19 @@ describe('Query Rule creation with trigger index and script action', () => {
 	});
 
 	it('Should open query rules page', () => {
-		cy.visit(`${base_url}/cluster/rules`).wait(PAGE_LOAD_TIME);
+		cy.server();
+		cy.route('/arc/plan').as('plan');
+		cy.route('**/_rules').as('rules');
+		cy.route('**/_aliasedindices').as('indices');
+		cy.visit(`${base_url}/cluster/rules`);
+		cy.wait(['@plan', '@rules', '@indices'], { timeout: 25000 });
 	});
 
-	it('Should navigate to create query rule page', () => {
+	it('Should open new query rule form page', () => {
+		cy.server();
+		cy.route('**/_mapping').as('mapping');
 		cy.get('[data-cy=create-query-rule]').click();
-		cy.wait(PAGE_LOAD_TIME);
+		cy.wait('@mapping', { timeout: 15000 }).wait(5000);
 	});
 
 	it('Should create a query rule', () => {
@@ -40,7 +50,7 @@ describe('Query Rule creation with trigger index and script action', () => {
 		cy.get('[name="description"]').type('cypress-testing-rule-description');
 
 		// Select index
-		cy.get('[data-cy=index-dropdown]').click().type('best');
+		cy.get('[data-cy=index-dropdown]').click().type('best-buy-data');
 		cy.get('[data-cy=best-buy-dataset]').click({ force: true, multiple: true });
 
 		// Select always trigger type
@@ -62,10 +72,13 @@ describe('Query Rule creation with trigger index and script action', () => {
 		}).as('save');
 
 		cy.get('[data-cy=save-query-rule]').click();
-		cy.wait('@save', { timeout: 15000 });
-
-		cy.get('@save').then((xhr) => {
-			const ruleId = xhr?.response?.body?.id || null;
+		cy.wait('@save', { timeout: 15000 }).then((xhr) => {
+			ruleId = xhr?.response?.body?.id || null;
+		});
+	});
+	it('Should delete query rule', () => {
+		const credentials = btoa(`${username}:${password}`);
+		if (ruleId) {
 			cy.request({
 				method: 'DELETE',
 				url: `${app_url}_rule/${ruleId}`,
@@ -73,8 +86,8 @@ describe('Query Rule creation with trigger index and script action', () => {
 					Authorization: `Basic ${credentials}`,
 				},
 			}).wait(2000);
-			cy.visit(`${base_url}/cluster/rules`);
-		});
+			cy.visit(`${base_url}/cluster/rules`).wait(PAGE_LOAD_TIME);
+		}
 	});
 	it('Should logout user', () => {
 		cy.clearLocalStorage();
