+++ cypress/integration/add-index-trigger-script-action.spec.js
@@ -1,5 +1,9 @@
 import generateName from '../utils/generateName';
 import { base_url, username, password, app_url, cluster } from '../utils/index';
+import { PAGE_LOAD_TIME } from './contants';
+
+// Track query rule id, to delete later
+let ruleId;
 
 describe('Query Rule creation with trigger index and script action', () => {
 	before(() => {
@@ -25,12 +29,22 @@ describe('Query Rule creation with trigger index and script action', () => {
 	});
 
 	it('Should open query rules page', () => {
-		cy.visit(`${base_url}/cluster/rules`).wait(2000);
+		cy.server();
+		cy.route('/arc/plan').as('plan');
+		cy.route('**/_rules').as('rules');
+		cy.route('**/_aliasedindices').as('indices');
+		cy.visit(`${base_url}/cluster/rules`);
+		cy.wait(['@plan', '@rules', '@indices'], { timeout: 25000 });
 	});
 
-	it('Should create a query rule', { retries: 3 }, () => {
+	it('Should open new query rule form page', () => {
+		cy.server();
+		cy.route('**/_mapping').as('mapping');
 		cy.get('[data-cy=create-query-rule]').click();
+		cy.wait('@mapping', { timeout: 15000 }).wait(5000);
+	});
 
+	it('Should create a query rule', () => {
 		// Enter name and description
 		cy.get('[name="name"]').type('cypress-testing-rule-name');
 		cy.get('[name="description"]').type('cypress-testing-rule-description');
@@ -41,7 +55,8 @@ describe('Query Rule creation with trigger index and script action', () => {
 		// Select script action
 		cy.get('[data-cy=query-rule-action]').click();
 		cy.get('[data-cy=script]').click({ force: true, multiple: true });
-		cy.wait(2000);
+		// Wait until execution context is loaded
+		cy.wait(10000);
 		// Remove tooltip blocking the select box
 		cy.get('.ant-tooltip').then((tooltipEl) => {
 			console.log({ tooltipEl });
@@ -50,13 +65,12 @@ describe('Query Rule creation with trigger index and script action', () => {
 		cy.wait(2000);
 		cy.get('[data-cy=script-template]').click().type('a');
 		cy.wait(2000);
-		cy.get('[data-cy=asyncFetch]').wait(1000).click();
-		cy.wait(2000);
+		cy.get('[data-cy=asyncFetch]').click().wait(5000);
+
 		cy.get('[data-cy=query-rule-save-script]').click();
 		cy.wait(2000);
 
 		// Save query rule
-		const credentials = btoa(`${username}:${password}`);
 		cy.server();
 		cy.route({
 			method: 'POST',
@@ -64,10 +78,13 @@ describe('Query Rule creation with trigger index and script action', () => {
 		}).as('save');
 
 		cy.get('[data-cy=save-query-rule]').click();
-		cy.wait('@save', { timeout: 15000 });
-
-		cy.get('@save').then((xhr) => {
-			const ruleId = xhr?.response?.body?.id || null;
+		cy.wait('@save', { timeout: 15000 }).then((xhr) => {
+			ruleId = xhr?.response?.body?.id || null;
+		});
+	});
+	it('Should delete query rule', () => {
+		const credentials = btoa(`${username}:${password}`);
+		if (ruleId) {
 			cy.request({
 				method: 'DELETE',
 				url: `${app_url}_rule/${ruleId}`,
@@ -75,8 +92,8 @@ describe('Query Rule creation with trigger index and script action', () => {
 					Authorization: `Basic ${credentials}`,
 				},
 			}).wait(2000);
-			cy.visit(`${base_url}/cluster/rules`);
-		});
+			cy.visit(`${base_url}/cluster/rules`).wait(PAGE_LOAD_TIME);
+		}
 	});
 	it('Should logout user', () => {
 		cy.clearLocalStorage();
